Fixed the WorkoutDatePicker.kt so it's displayed on the same row as the other buttons as well as moved it into it's corresponding package
