+++ app/src/main/java/com/example/gymsmart/components/pages/workouts/UserWorkoutsPage.kt
@@ -14,7 +14,7 @@ import android.util.Log
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.platform.LocalContext
 import androidx.navigation.NavController
-import com.example.gymsmart.components.WorkoutDatePicker
+import com.example.gymsmart.components.ui.WorkoutDatePicker
 import com.example.gymsmart.components.ui.FilterDropdownMenu
 import com.example.gymsmart.components.ui.SearchBarWithIcon
 import com.example.gymsmart.components.ui.UserSettingsDropdownMenu
@@ -31,7 +31,6 @@ import java.util.Locale
  * @param navController
  */
 @Composable
-// TODO: Figure out a way to take the com.example.gymsmart.firebase.FirebaseAuthHelper helper as an argument, since we need to pass down.
 fun UserWorkoutsPage(
     navController: NavController,
     firebaseAuthHelper: FirebaseAuthHelper) {
@@ -132,12 +131,14 @@ fun UserWorkoutsPage(
                 horizontalArrangement = Arrangement.SpaceBetween
             ) {
                 // User Settings Drop Down Menu Button
-                // TODO: Figure out a way to pass down the firebaseAuthHelper variable down to the UserSettingsDropdownMenu
                 UserSettingsDropdownMenu({ setting -> navigateUserSettingMenu(setting) }, firebaseAuthHelper, navController )
 
 
                 // Filter Button
                 FilterDropdownMenu { filter -> applyFilter(filter) }
+
+                // Calender Filter Button
+                WorkoutDatePicker(LocalContext.current) { newDate -> applyCalendarFilter(newDate) }
 
                 // Search Bar
                 SearchBarWithIcon(searchQuery) { query -> searchQuery = query }
@@ -207,7 +208,6 @@ fun UserWorkoutsPage(
                         }
                     }
                 }
-                WorkoutDatePicker(LocalContext.current, {newDate -> applyCalendarFilter(newDate) }, workouts)
             }
         }
     }
